Release version bumped to 1.5.
Added call to gen_cc.pl before packing to get country code list in the release.
List of the files moved into the variable, to make it consistent. Files gen_cc.pl
and list are now excluded from the release. Fixed problem of not including
Organizations/ into the zip archive.